+++ src/query.rs
@@ -2,3 +2,4 @@ pub mod album;
 pub mod artist;
 pub mod audiobook;
 pub mod category;
+pub mod show;